+++ problems/Problem_181.hpp
@@ -17,12 +17,15 @@ inline std::vector<std::string> splitPalindrome( std::string str )
 {
   std::vector<std::string> palindromes;
 
-  while(!str.empty()){
-    for(int i = str.size(); i > 0; i--){
-      std::string substr = str.substr(0,i);
-      if(isPalindrome(substr)){
-        palindromes.push_back(substr);
-        str = str.substr(i);
+  while( !str.empty() )
+  {
+    for( size_t i = str.size(); i > 0; i-- )
+    {
+      std::string substr = str.substr( 0, i );
+      if( isPalindrome( substr ) )
+      {
+        palindromes.push_back( substr );
+        str = str.substr( i );
         break;
       }
     }
